refactor: 상태관리 커스텀 훅으로 이동
