+++ src/pages/alba/ApplyPage/TimeSelectSection/DailySelect.tsx
@@ -2,54 +2,45 @@ import React, { useEffect } from 'react';
 import FlexContainer from 'components/@commons/FlexContainer';
 import BorderBox from 'components/@commons/BorderBox';
 import Text from 'components/@commons/Text';
-import { SelectedSchedule, putApply } from 'apis/alba/apply';
+import { SelectedSchedule } from 'apis/alba/apply';
 import CheckBox from 'components/@commons/CheckBox';
-import { useAtom } from 'jotai';
-import { weeklySelectAtom } from '..';
 import useApply from 'hooks/alba/useApply';
-import { useMutation } from '@tanstack/react-query';
+import SubmitButton from 'components/@commons/SubmitButton';
 
 const DailySelect = ({ day, startWeekDate }: { day: number; startWeekDate: string }): JSX.Element => {
-  const { query, findTimeData } = useApply(startWeekDate);
+  const { weeklySelect, findTimeData, selectHandler, previewHandler, putSaveHandler } = useApply(startWeekDate);
 
-  const [weeklySelect, setWeeklySelect] = useAtom(weeklySelectAtom);
-
-  // 체크 시간 업데이트
-  const selectHandler = (timeObject: SelectedSchedule, timeIndex: number) => {
-    const newDaily = weeklySelect[day].map((selected: SelectedSchedule, i) =>
-      i === timeIndex ? { ...timeObject, isChecked: !timeObject.isChecked } : selected,
-    );
-    setWeeklySelect((prevWeekly) => prevWeekly.map((prevDaily, dayIndex) => (dayIndex === day ? newDaily : prevDaily)));
-  };
-
-  // 요일 바뀔 때 마다 서버에 저장
-  const mutation = useMutation(putApply, { onSuccess: () => query.refetch() });
+  // 요일 바뀔 때마다 서버에 저장
   useEffect(() => {
-    mutation.mutate({ weekStartDate: startWeekDate, apply: weeklySelect });
+    if (!weeklySelect.some((e) => e.length > 0)) return;
+    putSaveHandler(day);
   }, [day]);
 
   return (
     <>
-      {weeklySelect[day].map((timeObject: SelectedSchedule, timeIndex) => (
-        <label key={timeObject.workTimeId}>
-          <BorderBox width="100%" gradation={true}>
-            <FlexContainer $wFull $padding="28px" $direction="row" $align="center">
-              <CheckBox
-                type="checkbox"
-                onClick={() => selectHandler(timeObject, timeIndex)}
-                checked={timeObject.isChecked}
-                readOnly
-              />
-              <Text size="xl" margin="0">
-                {findTimeData(timeObject.workTimeId).title}
-              </Text>
-              <Text size="xl" margin="0 0 0 auto">
-                {findTimeData(timeObject.workTimeId).startTime} ~ {findTimeData(timeObject.workTimeId).endTime}
-              </Text>
-            </FlexContainer>
-          </BorderBox>
-        </label>
-      ))}
+      <FlexContainer $wFull>
+        {weeklySelect[day].map((timeObject: SelectedSchedule, timeIndex) => (
+          <label key={timeObject.workTimeId}>
+            <BorderBox width="100%" gradation={true}>
+              <FlexContainer $wFull $padding="28px" $direction="row" $align="center">
+                <CheckBox
+                  type="checkbox"
+                  onClick={() => selectHandler(timeObject, timeIndex, day)}
+                  checked={timeObject.isChecked}
+                  readOnly
+                />
+                <Text size="xl" margin="0">
+                  {findTimeData(timeObject.workTimeId).title}
+                </Text>
+                <Text size="xl" margin="0 0 0 auto">
+                  {findTimeData(timeObject.workTimeId).startTime} ~ {findTimeData(timeObject.workTimeId).endTime}
+                </Text>
+              </FlexContainer>
+            </BorderBox>
+          </label>
+        ))}
+      </FlexContainer>
+      <SubmitButton onClick={previewHandler}>미리보기</SubmitButton>
     </>
   );
 };
